Move the fake server subsystem into src/core
