+++ src/core/fake/fake-io.js
