+++ src/core/fake/fake-socket.js
