+++ src/index.js
@@ -9,13 +9,13 @@ export { internal }
 export * from './types/types.js'
 
 // Ancillary exports:
+export { makeFakeIos } from './core/fake/fake-io.js'
+export { fakeUser } from './core/fake/fakeUser.js'
+export { fakeUser1 } from './core/fake/fakeUser1.js'
 export { destroyAllContexts } from './core/root.js'
 export { makeBrowserIo } from './io/browser/browser-io.js'
-export { makeFakeIos } from './io/fake/fake-io.js'
 export { makeNodeIo } from './io/node/node-io.js'
 export { makeReactNativeIo } from './io/react-native/react-native-dummy.js'
-export { fakeUser } from './io/fake/fakeUser.js'
-export { fakeUser1 } from './io/fake/fakeUser1.js'
 export {
   DustSpendError,
   errorNames,
