+++ src/makeContext.js
@@ -2,9 +2,9 @@
 
 import { isReactNative } from 'detect-bundler'
 
+import { prepareFakeIos } from './core/fake/fake-io.js'
 import { makeContext } from './core/root.js'
 import { makeBrowserIo } from './io/browser/browser-io.js'
-import { prepareFakeIos } from './io/fake/fake-io.js'
 import { isNode, makeNodeIo } from './io/node/node-io.js'
 import { makeReactNativeIo } from './io/react-native/react-native-dummy.js'
 import {
